change decimals to floats, optimize structures, make movements for both chairperson and reviewer/supervisor pair

+++ Optimizer/Optimizer.Logic/Work/Score/Heuristics/IHeuristic.cs
@@ -2,5 +2,5 @@
 
 public interface IHeuristic
 {
-    decimal CalculateScore(PartialSolution solution);
+    float CalculateScore(PartialSolution solution);
 }
